# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReplyX API",
    "description": "\nReplyX - Comprehensive AI Assistant Platform API\n\n## Features\n- 🤖 AI Assistant Management\n- 💬 Real-time Dialog System  \n- 📄 Document Processing & Knowledge Base\n- 🤖 Telegram Bot Integration\n- 💰 Balance & Billing System\n- 👥 User Management & Authentication\n- 📊 Analytics & Monitoring\n- 🛡️ Security & Rate Limiting\n\n## Authentication\nAll endpoints (except public ones) require JWT authentication:\n- Include `Authorization: Bearer <token>` header\n- Include `X-CSRF-Token` header for state-changing operations\n\n## Base URLs\n- Production: `https://api.replyx.ru`\n- Development: `http://localhost:8000`\n        ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.replyx.ru",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/knowledge/{knowledge_id}": {
      "put": {
        "summary": "Update Knowledge",
        "operationId": "update_knowledge_api_knowledge__knowledge_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "knowledge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Knowledge Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Knowledge",
        "description": "Улучшенное удаление записи знаний с полной очисткой всех связанных данных",
        "operationId": "delete_knowledge_api_knowledge__knowledge_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "knowledge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Knowledge Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assistants/{assistant_id}/knowledge": {
      "get": {
        "summary": "Get Assistant Knowledge",
        "description": "Получить знания конкретного ассистента",
        "operationId": "get_assistant_knowledge_api_assistants__assistant_id__knowledge_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Assistant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/support/send-to-admins": {
      "post": {
        "summary": "Send Message To Admins",
        "description": "Принимает сообщения от виджета и создает диалог для администраторов",
        "operationId": "send_message_to_admins_api_support_send_to_admins_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeIn": {
        "properties": {
          "doc_id": {
            "type": "integer",
            "title": "Doc Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type"
          },
          "importance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance",
            "default": 10
          },
          "assistant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistant Id"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "content",
          "type"
        ],
        "title": "KnowledgeIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}